Mirror and glass transmission
We use `reflect` in order to trace the mirror reflection.

We use `refract` in order to trace the glass refraction. But glass also
have a mirror component which is traced using `reflect`. In order to
avoid tracing both rays, and generating a combinatorial explosion, we
uses importance sampling and randomly pick one ray.

For now we use IO based `randomIO` number generator. This must be
improved in the future in order to keep the function pure.