Display runtime informations (time and memory)
This commits display total runtime as well an builtime for the
acceleration structure and runtime for the rendering part.

In order to ensure that the builditem of the acceleration structure is
not lazyly done during the rendering part, we use deepseq, hence the new
NFData and Generic instances everywhere.

Note that this may also improve performances a bit because threads won't
compete during rendering in order to build the data structure.