Performance: refactor of the scene hierarchy.
Before this commit, each primitive `Sphere` was associated with a
`Material` with the `Object` type. The direct result was that an
aggregate of N `Sphere` was also containing N `Material`.

This was not great for performances. In the future, big mesh with
thousands of polygon should be only associated with one material.

This commit do a huge refactor with the following changes:

- `Object` associate a `Material` to a container of primitive, such as a
  list, a vector or a `BVH` of primitive or objects.
- `BVH` will store many (10 right now) items in its `BVHLeaf`. This
  improves performances, up unto a point, the tree overhead does not
  bring benefit.
- `BVH` may now store raw Sphere, or Object.
- The intersections functions were a bit changed to accomodate for the
  new interface. Especially, we introduced an `Intersect` and
  `HasBoundingBox` type classes which generalize theses concepts. The
  'IsObjectOrNot` type family allows to know if the current intersected
  item is associated to a `Material` or not.
- `Intersection` is now parametric, in order to store primitive or
  `Object`.

On the scene `ManySpheres 5`, the buildtime of the BVH is dramatically
improved, as well as a small improvement of the rendering time
(before/after):

BVH time: 16.106600017s
Rendering time: 1.053272759s 6%
Total time: 17.159872776s
Max memory usage:  867 MiB

BVH time: 2.256579807s
Rendering time: 0.652566709s 22%
Total time: 2.909146516s
Max memory usage:  150 MiB

On the `ManySpheres 2` scene, with