Extended scene graph
`Scene` is changed so now it can stores a more complex datatype. Instead
of a `BVH` of `Object`, it can now stores a `SceneGraph` which can
represents combinations of multiples `SceneGraph`. This will allows
later for extension, such as scene graph transformations.

`AttachMaterial` nows behaves as what `Object` was doing previously and
attach a `Material` to a `Geometry`. This `Geometry` can be `BVH` of
different primitives, such as `Sphere`, `Box` and later `Triangles`.

The `ManySphere` scene evolved to now show `Box` as well as `Sphere`.