Multiple importance sampling
This commit is not super clean. It was committed a few days after the
stream which introduces multiple importance sampling.

The MIS scene now uses two strategies to compute light, by sampling the
light source or when a ray hit an emitting surface and weights it using
multiple importance sampling.